스케줄 생성
스케줄 전체 조회
스케줄 상세 조회
스케줄 수정
스케줄 삭제

+++ services/schedule.service.js
@@ -1,24 +1,57 @@
-import { ScheduleRepository } from '../repositories';
+import { ScheduleRepository, UserRepository, TrainerRepository } from '../repositories';
 class ScheduleService {
   _scheduleRepository = new ScheduleRepository();
+  _userRepository = new UserRepository();
+  _trainerRepository = new TrainerRepository();
 
-  postSchedule = async (title, date) => {
+  // 스케줄 생성
+  postSchedule = async (title, date, userId) => {
     try {
       if (!title) {
         return {
           status: 400,
           message: '제목 미입력',
         };
-      } else if (!date) {
+      }
+
+      if (!date) {
         return {
           status: 400,
           message: '날짜 미입력',
         };
       }
-      const result = await this._scheduleRepository.postSchedule(title, date);
+
+      const user = await this._userRepository.getUserById(userId);
+
+      if (user.type !== 'trainer') {
+        return {
+          status: 401,
+          message: '스케줄 생성 권한 없음',
+        };
+      }
+
+      const trainer = await this._trainerRepository.findTrainer(userId);
+
+      if (!trainer) {
+        return {
+          status: 404,
+          message: '트레이너 정보 없음',
+        };
+      }
+
+      const checkDate = await this._scheduleRepository.checkDateSchedule(date);
+
+      if (checkDate) {
+        return {
+          status: 409,
+          message: '중복된 날짜',
+        };
+      }
+
+      const result = await this._scheduleRepository.postSchedule(title, date, userId, trainer.trainerId);
       if (!result) {
         return {
-          status: 400,
+          status: 404,
           message: '업체 생성 실패',
         };
       }
@@ -31,7 +64,22 @@ class ScheduleService {
     }
   };
 
-  oneGetSchedule = async scheduleId => {
+  // 스케줄 전체 조회
+  allGetSchedule = async userId => {
+    try {
+      const result = await this._scheduleRepository.allGetSchedule(userId);
+
+      return {
+        status: 200,
+        message: result,
+      };
+    } catch (err) {
+      return { status: 500, message: 'Server Error' };
+    }
+  };
+
+  // 스케줄 상세 조회
+  oneGetSchedule = async (scheduleId, userId) => {
     try {
       if (!scheduleId) {
         return {
@@ -39,10 +87,20 @@ class ScheduleService {
           message: '스케줄 ID 미입력',
         };
       }
+
+      const user = await this._userRepository.getUserById(userId);
+
+      if (user.type !== 'trainer' || user.type !== 'admin') {
+        return {
+          status: 401,
+          message: '스케줄 조회 권한 없음',
+        };
+      }
+
       const result = await this._scheduleRepository.oneGetSchedule(scheduleId);
       if (!result) {
         return {
-          status: 400,
+          status: 404,
           message: '존재하지 않는 스케줄 ID',
         };
       }
@@ -55,7 +113,8 @@ class ScheduleService {
     }
   };
 
-  putSchedule = async (scheduleId, title, date) => {
+  // 스케줄 수정
+  putSchedule = async (scheduleId, title, date, userId) => {
     try {
       if (!scheduleId) {
         return {
@@ -63,27 +122,42 @@ class ScheduleService {
           message: '스케줄 ID 미입력',
         };
       }
+
       const schedule = await this._scheduleRepository.oneGetSchedule(scheduleId);
       if (!schedule) {
         return {
-          status: 400,
+          status: 404,
           message: '존재하지 않는 스케줄 ID',
         };
-      } else if (!title) {
+      }
+
+      const user = await this._userRepository.getUserById(userId);
+
+      if (schedule.userId !== user.userId || user.type !== 'admin') {
+        return {
+          status: 401,
+          message: '스케줄 수정 권한 없음',
+        };
+      }
+
+      if (!title) {
         return {
           status: 400,
           message: '제목 미입력',
         };
-      } else if (!date) {
+      }
+
+      if (!date) {
         return {
           status: 400,
           message: '날짜 미입력',
         };
       }
+
       const result = await this._scheduleRepository.putSchedule(scheduleId, title, date);
       if (!result) {
         return {
-          status: 400,
+          status: 404,
           message: '스케줄 수정 실패',
         };
       }
@@ -95,7 +169,9 @@ class ScheduleService {
       return { status: 500, message: 'Server Error' };
     }
   };
-  deleteSchedule = async scheduleId => {
+
+  // 스케줄 삭제
+  deleteSchedule = async (scheduleId, userId) => {
     try {
       if (!scheduleId) {
         return {
@@ -103,17 +179,28 @@ class ScheduleService {
           message: '스케줄 ID 미입력',
         };
       }
+
       const schedule = await this._scheduleRepository.oneGetSchedule(scheduleId);
       if (!schedule) {
         return {
-          status: 400,
+          status: 404,
           message: '존재하지 않는 스케줄 ID',
         };
       }
+
+      const user = await this._userRepository.getUserById(userId);
+
+      if (schedule.userId !== user.userId || user.type !== 'admin') {
+        return {
+          status: 401,
+          message: '스케줄 삭제 권한 없음',
+        };
+      }
+
       const result = await this._scheduleRepository.deleteSchedule(scheduleId);
       if (!result) {
         return {
-          status: 400,
+          status: 404,
           message: '스케줄 삭제 실패',
         };
       }
